fix(attendancereport): round up to int on attendance percentage

+++ src/api/Libraries/Modules/ECDLink.SmartStart/Reports/Models/ChildAttendanceMonthlyReportModel.cs
@@ -10,6 +10,6 @@ namespace ECDLink.SmartStart.Reports.Models
 
         public int ExpectedAttendance { get; set; }
 
-        public decimal AttendancePercentage { get; set; }
+        public int AttendancePercentage { get; set; }
     }
 }
