+++ src/api/Libraries/Modules/ECDLink.SmartStart/Reports/Models/ChildGroupingAttendanceReportModel.cs
@@ -17,7 +17,7 @@ namespace ECDLink.SmartStart.Reports.Models
 
         public int ActualAttendance { get; set; }
 
-        public decimal AttendancePercentage { get; set; }
+        public int AttendancePercentage { get; set; }
 
         public IEnumerable<ChildAttendanceMonthlyReportModel> MonthlyAttendance { get; set; }
     }
